Improved tests for JwtBuilder (#260)
* Improved tests for JwtBuilder
* Simplified linking tests from Common
* Made all tests having same namespace
