+++ tests/JWT.Tests.Common/JWT.Tests.Common.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
+    <RootNamespace>JWT.Tests</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
