+++ tests/JWT.Tests.Common/JwtEncoderTests.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using FluentAssertions;
 using JWT.Algorithms;
 using JWT.Serializers;
-using JWT.Tests.Common.Models;
+using JWT.Tests.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace JWT.Tests.Common
